# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insider Messaging API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Retrieve messages",
                "description": "Fetch messages with optional filters such as limit and sent status.",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of messages to retrieve",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "isSent",
                        "in": "query",
                        "description": "Filter by sent status (1 for sent, 0 for unsent)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Messages fetched successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "67f2f8a8-ea58-4ed0-a6f9-ff217df4d849"
                                            }
                                        },
                                        "isSuccess": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid parameters"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "isSuccess": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Messages",
            "description": "Messages"
        }
    ]
}